+++ lib/widgets/scaffold/basic_scaffold.dart
@@ -15,11 +15,19 @@ class BasicScaffold extends StatefulWidget {
   /// Property contains a [drawer] which is located on the appBar
   Widget? drawer;
 
+  /// Property contains a bool which shows a [floatingActionButton]
+  bool? showButton;
+
+  /// Property contains a [onButtonPressed] callback whenever the button is pressed
+  Function()? onButtonPressed;
+
   BasicScaffold({
     required this.title,
     required this.body,
     this.actions,
     this.drawer,
+    this.showButton,
+    this.onButtonPressed,
   }) : super();
 
   @override
@@ -36,6 +44,16 @@ class _BasicScaffoldState extends State<BasicScaffold> {
         actions: widget.actions,
       ),
       body: widget.body,
+      floatingActionButton: (widget.showButton ?? false)
+          ? FloatingActionButton(
+              elevation: 10.0,
+              onPressed: widget.onButtonPressed,
+              child: Icon(
+                Icons.add,
+                color: Theme.of(context).iconTheme.color,
+              ),
+            )
+          : null,
     );
   }
 }